app/src/requirements.txt
Flask~=3.0.3
pymongo~=4.8.0
bcrypt~=4.0.1
PyJWT~=2.7.0
Flask-CORS~=3.0.10
joblib~=1.2.0
paho-mqtt~=1.6.1
numpy~=1.24.0
pandas~=1.5.2
requests~=2.28.1
scikit-learn~=1.5.1